I modified tl_detector.py aiming to alleviate the latency issue. The key idea is to adjust the frequency of Clf calls adaptively.
if the distance between the car and the closest traffic light is larger than DIFF_THRESHOLD
Clf will be applited every 8th images.
otherwise, it's 4th.

I tested it on my local machine (w/o GPU).
Without this modification, the latency increases dramatically at the end of the lap. (same as what Jeremy observed)
With this modification, the car can successfully make it for one complete lap.